+++ pkg/services/auth/authimpl/token_cleanup.go
@@ -1,4 +1,4 @@
-package auth
+package authimpl
 
 import (
 	"context"
