+++ pkg/services/auth/authimpl/token_cleanup_test.go
@@ -1,4 +1,4 @@
-package auth
+package authimpl
 
 import (
 	"context"
